Lumen Authorization

+++ app/Http/Controllers/PostsController.php
@@ -5,12 +5,26 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostsController extends Controller
 {
     public function index()
     {
-        $posts = Post::where(['user_id' => Auth::user()->id])->orderBy("id", "DESC")->paginate(2)->toArray();
+        if (Gate::denies('read-post')) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You Are unauthorized'
+            ], 403);
+        }
+
+        if (Auth::user()->role === 'admin') {
+            $posts = Post::OrderBy("id", "DESC")->paginate(2)->toArray();
+        } else {
+            $posts = Post::where(['user_id' => Auth::user()->id])->OrderBy("id", "DESC")->paginate(2)->toArray();
+        }
+
         $response = [
             "total_count" => $posts["total"],
             "limit" => $posts["per_page"],
@@ -22,20 +36,28 @@ class PostsController extends Controller
         ];
 
         return response()->json($response, 200);
-
-        // $outPut = [
-        //     "message" => "posts",
-        //     "results" => $posts
-
-        // ];
-
-        // return response()->json($posts, 200);
     }
 
     public function store(Request $request)
     {
-        $input = $request->all();
+        $input = [
+            'title' => $request->input('title'),
+            'content' => $request->input('content'),
+            'status' => $request->input('status'),
+            'user_id' => Auth::user()->id
+        ];
+
+        if (Gate::denies('store-post')) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You Are unauthorized'
+            ], 403);
+        }
+
         $post = Post::create($input);
+
+
 
         return response()->json($post, 200);
     }
@@ -47,6 +69,27 @@ class PostsController extends Controller
         if (!$post) {
             abort(404);
         }
+
+        if (Gate::denies('read-post', $post)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You Are unauthorized'
+            ], 403);
+        }
+
+        if (Auth::user()->role === 'admin') {
+            return response()->json($post, 200);
+        }
+
+        if (Auth::user()->role === 'editor' && $post->user_id !== Auth::user()->id) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You Are unauthorized'
+            ], 403);
+        }
+
         return response()->json($post, 200);
     }
 
@@ -58,6 +101,26 @@ class PostsController extends Controller
 
         if (!$post) {
             abort(404);
+        }
+
+        if (Gate::denies('update-post', $post)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized'
+            ], 403);
+        }
+
+        $validationRules = [
+            'title' => 'required|min:5',
+            'content' => 'required|min:10',
+            'status' => 'required|in:draft,published',
+        ];
+
+        $validator = \Validator::make($input, $validationRules);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
         }
 
         $post->fill($input);
@@ -73,6 +136,14 @@ class PostsController extends Controller
 
         if (!$post) {
             abort(404);
+        }
+
+        if (Gate::denies('update-post', $post)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized'
+            ], 403);
         }
 
         $post->delete();
